\name{table}
\docType{data}
\alias{cutval}
\alias{table}

\title{Pvalue Data Tables}

\description{
  Stores cutoff information for different target alpha values and
  various sets of data of varying sample size. 
}

\format{
  data.frame with columns equal to sample size information
  and rows equal to different target alpha values.
}

\details{
  This file contains cutoff information for different target alpha values and
  various sets of data of varying sample size. This table is generated for sample sizes 2-30, 35, 40, 45, 50, 55, 60, 70, 80, 90, 100, 120, 150, 170, 200, 225, 250, 275, 300, 325, 350, 375, 400, 425, 450. The target alphas range
  from .1 to .9 in increments of .1. The delta is 0.1
}

\note{
  This dataset is used within the tsc.test function. There is no
  need for the user to ever call this dataset.
}

\keyword{datasets}





